feat: Improve the FileSystemTestCase (#57)
Simplify the usage by no longer requiring to implement an abstract
method.

A significant (internal, but observable) change is that we no longer use
a namespaced directory. Indeed, instead of having:

```
/path/to/system/default-tmp/
  TestCaseThreadSafeNamespace/
    tmpDir1
    tmpDir2
    ...
```

We now have:

```
/path/to/system/default-tmp/
  tmpDir1
  tmpDir2
  ...
```

Indeed, the creation of the temporary directory is already thread safe,
so there is no need for the user to implement an abstract method. And
because of it, we can remove the usage of a namespace which is not that
useful and if anything may mess up the thread safety.